requirements.txt
streamlit
crawl4ai
python-dotenv
aiohttp
google-generativeai